If we add the ability to store failures, we can isolate this failure and report it in our own tests in a meaningful way.
We can now see the fundamental issue in our code - the setup and should blocks are run as soon as they are encountered, but we need to at the very least defer running the setup until the should blocks run, and then run the setup before each one. That's next.