Failing test for running subcontext tests only
OK, that was simple so far. Now I want to do things a bit differently - I want to be able to programmatically grab a subcontext and run *only* the tests within it. With test/unit, you need to do with with regexp and stuff from the command line, but I don't see why these things can't be objects, and why they can't be run by invoking methods.