Code after a failing assertion shouldn't be run.
This means we need to escape the flow of control of the test, which forces us to use exceptions for the assertion mechanism. I don't think this is all bad though; it also means we can remove the instance variable in the TestEnvironment, which could've been interfered with by a running test.