It's also idomatic to include methods from modules.
In this case, since the block is evaluated in the context of an instance, we actually want to extend the instance with that module to add the methods to the instance.

It occurs to me know that the method definition we previously added might be global for *all* contexts, which is possibly not desirable. Something to investigate, then.