We should be able to run tests by just running the file.
This is a bit convoluted, but basically I'm detecting whether or not the file that was run was the file that required jtest. Ish.

In order to make the jtest tests actually work, I need a way to opt-out of this behaviour. That's what the environment flag is about. A bit ugly, but it works.