Another way of defining what-would-be private methods in a Test::Unit testcase.
It's potentially confusing to be able to define methods in the context which can then be called in that context (i.e. 'class' methods), but also be able to define or include methods that can be called within tests or setups (i.e. 'instance' methods). However, I can't think of a better way around this right now.