Extracted tests into smaller groupings of related functionality for clarity.
I deliberately gave the superclass a horrible name - `JTest_TestUnit_TestCase` - because I didn't want it to seem like it was part of the library that I'm writing. Well, it made sense to me at the time.

I also had to make the subcontext ordering alphabetical to fix a test that would fail when rake was run.