A failing test demonstrating an issue defining methods that add tests.
The gist of the issue is that when defining a method in a context, the value of `self` when that method is executed will always be the context where the method was defined. So, when calling this method in a subcontext to try and add a test, the test will actually be defined in the parent context.

It's relatively simple to work around this by capturing method definitions in a module (whose methods are not tied to a particular value of `self`), but it's much harder to do this for bare method definitions.

In some ways it feels like this would be easy to solve if contexts were defined as classes, with subcontexts defined as subclasses, since this would clearly support the behaviour that's broken here. However, I'm not sure if moving to a class-based implementation would introduce other issues.