Re-implementing Kintama using classes and inheritance to solve the 'self in method' issue in the previous commit.
This is the simplest way of replicating the behaviour one would expect from `Test::Unit`-based frameworks. However, it does introduce some complications.

The key to understanding this is considering the value of self when a test runs, combined with where those methods are defined. For example:

    context "A context" do
      def a_method
        @thing
      end
      context "with a subcontext" do
        should "work" do
          a_method
        end
      end
    end

In order for this to behave as expected, the receiver for 'a_method' needs to be the same value as 'self' within the `should "work"` test. There is no simple way to intercept the definition of `a_method` into a module that can be included, whilst still supporting subcontexts and the other tangible behaviour that needs to be available during the definition of a context. I did look into `UnboundMethod`, but it can only be bound back to something of the same class.

The complication is that a Context is a class, but a Test needs to be an instance. It's actually an instance of a subclass of the Context, which is itself a subclass of any parent context. This is how the test body gains access to methods defined in the context while still retaining the correct value of `self` when those methods are invoked.

Having tests as instances makes the method-esque and hash-esque test running (as seen in `test_and_subcontext_access_test.rb`) a bit dirtier, since in order to inspect a test after it has been individually run, you need to retain the instance that was run. However, I'm not convinced about this functionality anyway, so am happy to overlook it.

My original concern with using class-based composition was that it would make the implementation less flexible, but I am no longer sure that this concern was founded.