The TestEnvironment class is no longer required.
The purpose of this class was to provide a blank-slate environment for tests to actually run, to make it simpler to inspect them after failures. We could still implement something like this by creating a final subclass of the test without any methods defined beyond those in Runnable already, but there are no tests that demand this so we won't.