Get kintama running on 1.9.2 and prep for rubinius.
Neither 1.9.2 nor rubinius exhibit the caller behaviour that allowed determining the originating method call in the same way as 1.8.7. The good news is that they have far, far better mechanisms. The bad news is that I've had to submit a patch to Rubinius to add source_location to Proc.

I also found a weird merge in the line running tests.